(service) housekeeping
Split gRPC implementation into multiple .cpp files to reduce the lines
of code in a single file.
